progress checkpoint: code might be broken
